Adds file inspection callback and example code
libclamav callbacks can be used to access embedded file content at each
layer of extraction during the course of a scan. The existing callbacks
only provide access to the file descriptor and a guess at the file type.

This patch adds a new callback for the purposes of file/archive
inspection that provides additional insight into the embedded file. This
includes:
- intermediate files: a ` > ` separated list of parent file names
- parent file size
- file name
- file size
- file buffer (pointer)
- type (just a guess)
- file descriptor (serves same purpose as buffer, but a different
interface)

Two new example apps are added that are automatically built when
compiling under CMake:
- ex2 demonstrates the prescan callback.
- ex3 demonstrates the new file inspection callback.

The examples are now installed if enabled, so you can test them in the
Docker image, and so that they'll be colocated with the DLLs so you can
test them on Windows. The installed examples should also be able to find
the UnRAR library at run time, without having to set LD_LIBRARY_PATH.

This commit also sets the fmap->name in an fmap-scan using the basname
of the provided filename if the caller provided the filename and the
provided fmap does not have the name set.

This commit adds logic to record if a files extracted from a ZIP was
decrypted and to present that to the file inspection callback function.