CAB/CHM: Remove allmatch checks & clean up error handling
Removed the allmatch checks that are no longer necessary now that it is
done when appending alerts.

Cleaned up error handling in the process to use the `goto done;` method.
I believe this fixes a minor bug wherein a detection in CAB or CHM
metadata scanning would end the scan early in allmatch mode.