Matcher: code cleanup, fix possible leaks
Added inline documentation and did some general cleanup of the
`cli_scan_buff`, and then updated the function comment now that I
understand the function a little better.

While doing this, I found that the calls to cli_ac_initdata were being
done regardless of whether or not logically initialized matcher data was
required or used.  But the call to free that matcher data was only being
done when AC-data was not provided by the caller.  This would be a leak.
I fixed this by only initalizing the AC data when AC data is not
provided by the caller.