Further reduce AC pattern memory usage
The AC pattern store both a virname and a customdata field, and uses the
two in a mutually exclusive way, with customdata only used with virname
NULL and vice versa.
The customdata field is used to store a list of regexes that would be
run on the buffer if the AC pattern matched. Specifically it is used
only for domain names for the phishing module.

This commit seeks to unify the two.
This commit is mostly complete, but has one flaw, it improperly free's
the regex struct.

Additionally:
I am interested to know if we further minimize the AC pattern structure
by removing the virname/customdata/match_context variable entirely.
My idea is this: right now it stores an lsig ID, subsig ID, and such.
If we could somehow much those concepts with the pointer to the
match_contenxt, then perhaps we could look up the match_context /
virname / customdata when a match occurs, and not store it in this small
pattern structure at all. In the past, this AC pattern constituted a
whole signature, so it made sense to put the name here.  But now it is
often just one part of a signature (a sub-pattern in a larger
subsignature which may be in a sig that has multiple subsignatures).
Storing the name is wasteful.