clamonacc: fix build issue with static lib dependencies
When compiling clamonacc where libcurl and openssl libraries are static
libraries, you may encounter the following error:

  undefined reference to symbol 'dlclose@@GLIBC_2.2.5'

This implies that clamonacc needs to be linked with libdl.so.

To fix this, we link 'clamonacc' with `CMAKE_DL_LIBS` which is
predefined by CMake to point to whichever library has dlopen and
dlclose.

Ref: https://cmake.org/cmake/help/latest/variable/CMAKE_DL_LIBS.html
Ref: https://stackoverflow.com/questions/56841870/how-to-fix-undefined-reference-to-symbol-dlcloseglibc-2-2-5-from-glad-c