Record unique entity-id for each layer scanned
Every time we push a new map onto the scanning recursion context, give
it a unique entity id number, which counts from zero.

Moved the location where we add metadata for each file from the
"cli_magic_scan" function over to the "recursion stack push" function.

Include a "path" as a parameter for creating a new fmap, and rename some
related variables and functions to be more intuitive.