WIP: upgrade ac, bm, pcre, and hash size ints from uint32_t to size_t
This is a large WIP that mostly completes an upgrade of clamav's file
size integers for the core matcher modules from 32bits to `size_t` (aka
64bit for 64bit platforms).

Where I got stuck: I started upgrading the size limits for the bytecode
context structure and found the bytecode API is limited to 32bits witha
sprinkle of `off_t` (like ... the worst integer type).
I don't have the time or energy to try to fix this and it might never
work if it turns out the bytecode runtimes or signature language are
limited to 32bit).  If that's the case, we'd need to stick to 32bit
limits for the bytecode module and check file sizes against those
limites before doing anything bytecode related (aka "bytecode module
won't work for large files" typa thing).
It's very doable but I've sunk about 12 hours into this and it's a
stretch from my original goal of "64bit support for #-bytes scanned"
which I already completed in a previous commit.

So I'm setting this aside for now.