WIP: clamav.h: extensions to hashing functions with flags
Extensions to the cl_hash*() functions adding a flags parameter to indicate if FIPS should be ignored as well as if the function should allocate the new hash or use an existing buffer

Started work on the implementation which uses the new OpenSSL 3 mechanism for bypassing FIPS algorithm restrictions. The old implementation bypassed FIPS with the OpenSSL 1.x method that is no longer effective.

Note: we shouldn't bypass FIPS for the FP signatures or for verifying the legacy RSA-based CVD digital signatures.